% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedDipole.R
\name{draw_loaded_dipole_schematic}
\alias{draw_loaded_dipole_schematic}
\title{Draw a dipole schematic to PNG for a loaded dipole design}
\usage{
draw_loaded_dipole_schematic(design, file, width = 9, height = 3)
}
\arguments{
\item{design}{The output list from \code{design_loaded_dipole_full()}.}

\item{file}{Output PNG file path.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}
}
\value{
Invisibly, the file path.
}
\description{
Draw a dipole schematic to PNG for a loaded dipole design
}
\keyword{internal}
