% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedDipole.R
\name{draw_loaded_dipole_schematic_base64}
\alias{draw_loaded_dipole_schematic_base64}
\title{Draw a dipole schematic and return a Base64 string}
\usage{
draw_loaded_dipole_schematic_base64(design, width = 9, height = 3)
}
\arguments{
\item{design}{Design object from \code{design_loaded_dipole_full()}.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}
}
\value{
Base64-encoded PNG suitable for HTML embedding.
}
\description{
Draw a dipole schematic and return a Base64 string
}
\keyword{internal}
