% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedDipole.R
\name{plot_coil_turns}
\alias{plot_coil_turns}
\title{Plot inductance vs turns for a given PVC coil form}
\usage{
plot_coil_turns(
  pvc_size = "1-1/2",
  wire_diameter = 0.064,
  turn_spacing = NULL,
  N_range = 5:150
)
}
\arguments{
\item{pvc_size}{PVC nominal size.}

\item{wire_diameter}{Wire diameter in inches.}

\item{turn_spacing}{Turn spacing in inches.}

\item{N_range}{Range of turns to evaluate.}
}
\description{
Plot inductance vs turns for a given PVC coil form
}
