% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedDipole.R
\name{print.loadedDipoleRecommendation}
\alias{print.loadedDipoleRecommendation}
\title{Print method for loaded dipole recommendations}
\usage{
\method{print}{loadedDipoleRecommendation}(x, ...)
}
\arguments{
\item{x}{An object returned by \code{recommend_coil_position()}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for loaded dipole recommendations
}
