% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadedDipole.R
\name{print.loadedDipoleSolution}
\alias{print.loadedDipoleSolution}
\title{Print method for loaded dipole solutions}
\usage{
\method{print}{loadedDipoleSolution}(x, ...)
}
\arguments{
\item{x}{An object of class `loadedDipoleSolution` returned by
`solve_loaded_dipole`.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for loaded dipole solutions
}
